Running application and tables created, to add requests

+++ hotel-management-system-backend/src/main/java/com/codegen/hotelmanagementsystembackend/repository/CustomerRepository.java
@@ -3,5 +3,6 @@ package com.codegen.hotelmanagementsystembackend.repository;
 import com.codegen.hotelmanagementsystembackend.entities.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+
 }
